requirements.txt
pbr>=0.5.6
PyYAML
requests
six
